Small improvements, removed txt file duplicates
Skipped the first screenshot for every pose because of a problem whereby
the poses don't technically 'exist' to the UI at least because the
locations only change on render, making the green points look off. So
hence the skip. Files now have their own names and less confusing. No
need to divide each part of a vector in half it does that automatically.